Unique time series

Attempts to identify the unique time series that exist in this dataset.

A time series is defined as a unique combination of:

granularity_time

granularity_geo

country_iso3

location_code

border

age

sex

*_id

*_tag

unique_time_series(x, set_time_series_id = FALSE, ...)

Arguments

x

An object of type csfmt_rts_data_v2

set_time_series_id

If TRUE, then x will have a new column called 'time_series_id'

...

Not used.

Value

data.table, a dataset that lists all the unique time series in x.

See also